DeepDSL-Python-main/src/requirements.txt
antlr4-python3-runtime~=4.13.2
pip~=23.2.1
pillow~=11.1.0
matplotlib~=3.10.0
networkx~=3.4.2